perf: 하나의 컨트롤러에서 api 기능을 수행 하도록 변경
[api컨트롤러가 너무 많은듯 ㅇㅇ]
