--- dobby_music/src/main/java/com/project/side/music/controller/MusicAlbumConstoller.java
@@ -1,14 +1,0 @@
-package com.project.side.music.controller;
-
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/Album")
-@CrossOrigin(origins = "http://localhost:5174")
-public class MusicAlbumConstoller {
-}
